requirements-dev.txt
# Dev dependencies
-r requirements.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
coverage==7.3.2

# Code quality
black==25.1.0
isort==6.0.1          # ← Обновлено, чтобы совпадало с продакшн
flake8==7.2.0         # ← Обновлено, чтобы совпадало с продакшн
mypy==1.7.1
bandit==1.8.3         # ← Обновлено, чтобы совпадало с продакшн
safety

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.17.2
psycopg==3.1.18
